fix: remove WithProvisioningFolderMetadata from shared server
The flag caused move/delete jobs to produce warnings about folder
metadata and changed authorization for instance-scoped repos.

Only 3 tests need the flag (FixFolderMetadataJob, its auth test, and
JobsAuthorization). These are moved to jobs/foldermetadata/ with their
own shared server that has the flag enabled.

Made-with: Cursor
