--- pkg/tests/apis/provisioning/jobs/fixfoldermetadatajob_auth_test.go
@@ -1,82 +1,0 @@
-package jobs
-
-import (
-	"context"
-	"net/http"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-
-	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
-	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-)
-
-func TestIntegrationProvisioning_FixFolderMetadataJobAuthorization(t *testing.T) {
-	helper := sharedHelper(t)
-	ctx := context.Background()
-
-	const repo = "fixfoldermetadata-auth-test"
-	testRepo := common.TestRepo{
-		Name:               repo,
-		Target:             "folder",
-		Copies:             map[string]string{},
-		ExpectedDashboards: 0,
-		ExpectedFolders:    1,
-	}
-	helper.CreateRepo(t, testRepo)
-
-	body := common.AsJSON(provisioning.JobSpec{
-		Action: provisioning.JobActionFixFolderMetadata,
-	})
-
-	t.Run("admin can create fixFolderMetadata job", func(t *testing.T) {
-		var statusCode int
-		result := helper.AdminREST.Post().
-			Namespace("default").
-			Resource("repositories").
-			Name(repo).
-			SubResource("jobs").
-			Body(body).
-			SetHeader("Content-Type", "application/json").
-			Do(ctx).StatusCode(&statusCode)
-
-		require.NoError(t, result.Error(), "admin should be able to create fixFolderMetadata job")
-		require.Equal(t, http.StatusAccepted, statusCode, "should return 202 Accepted")
-
-		helper.AwaitJobs(t, repo)
-	})
-
-	t.Run("editor can create fixFolderMetadata job", func(t *testing.T) {
-		var statusCode int
-		result := helper.EditorREST.Post().
-			Namespace("default").
-			Resource("repositories").
-			Name(repo).
-			SubResource("jobs").
-			Body(body).
-			SetHeader("Content-Type", "application/json").
-			Do(ctx).StatusCode(&statusCode)
-
-		require.NoError(t, result.Error(), "editor should be able to create fixFolderMetadata job")
-		require.Equal(t, http.StatusAccepted, statusCode, "should return 202 Accepted")
-
-		helper.AwaitJobs(t, repo)
-	})
-
-	t.Run("viewer cannot create fixFolderMetadata job", func(t *testing.T) {
-		var statusCode int
-		result := helper.ViewerREST.Post().
-			Namespace("default").
-			Resource("repositories").
-			Name(repo).
-			SubResource("jobs").
-			Body(body).
-			SetHeader("Content-Type", "application/json").
-			Do(ctx).StatusCode(&statusCode)
-
-		require.Error(t, result.Error(), "viewer should not be able to create fixFolderMetadata job")
-		require.Equal(t, http.StatusForbidden, statusCode, "should return 403 Forbidden")
-		require.True(t, apierrors.IsForbidden(result.Error()), "error should be forbidden")
-	})
-}
