+++ pkg/tests/apis/provisioning/jobs/helpers_test.go
@@ -8,9 +8,7 @@ import (
 	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 )
 
-var env = common.NewSharedEnv(
-	common.WithProvisioningFolderMetadata,
-)
+var env = common.NewSharedEnv()
 
 func sharedHelper(t *testing.T) *common.ProvisioningTestHelper {
 	t.Helper()
